Create BYTEBEAT.c
This is a BYTEBEAT. You compile it with gcc -o mus BYTEBEAT.c and run it with ./mus>/dev/dsp in Debian stable (bookworm, 12).

The file is an algorithm that produces ones and zeros (bits) as integers into stdout with the putchar() C library function, that is included with C, the headers are #include <stdio.h>

The operators are bitwise XOR, bitwise AND, shift bits left, shift bits right and bitwise OR. It is easy to change the output stream of bits by changing the integers.

The copyright belongs to the original author, this is just my modification. I think "viznut" or someone else created the original.